consolidate cli into it's own module
make module-running a seperate concern
